requirements.txt
solana>=0.30.2
solders
anchorpy
borsh_construct
uuid